Move labelling utils into config.py
It's all just config access anyway
